rename-and-layout-changed
